Fix broken regex patterns, bug in constructor (#62)
I've fixed the error in the cog's regex patterns. The patterns are also simpler now, so they may not catch all edge cases, but it's a fix for now.
Also, I've fixed a logical bug we had in the constructor that wasn't affecting the bot: `self.bot` was assigned the value `Bot`, rather than `bot`.