Check if word is plural instead of using "(s)"
Before, the bot was not checking if the word 'month' or 'day' needed to be pluralised, so it would append "(s)" to the end of the word to account for both cases. However, this makes the bot seem less human, because people tailor their messages depending on the pluralisation.

Given that we needed to check for pluralisation in two places, and use the same suffix in both cases, it seemed intuitive to create a function which we could invoke twice. As I am unfamiliar with the codebase, it would be helpful to know if this should perhaps be defined in the global scope; or even in a file designed just for utility functions, because I could imagine this function being quite reusable.

This commit was made using the online interface due to me being away from my normal development computer. If there are formatting issues, I will correct them in a subsequent commit.