Disable the 'months' countdown when it's 0
When there are no months left, just say how many days are left.

There may be some efficiency concerns as we calculate the months countdown just to disable it! These are likely negligable. In the far future, we might want to consider caching the response every day -- but only if it's worth it.

This code has (still) not been tested.