Check if word is plural instead of using "(s)" (#83)
* Check if word is plural instead of using "(s)"

Before, the bot was not checking if the word 'month' or 'day' needed to be pluralised, so it would append "(s)" to the end of the word to account for both cases. However, this makes the bot seem less human, because people tailor their messages depending on the pluralisation.

Given that we needed to check for pluralisation in two places, and use the same suffix in both cases, it seemed intuitive to create a function which we could invoke twice. As I am unfamiliar with the codebase, it would be helpful to know if this should perhaps be defined in the global scope; or even in a file designed just for utility functions, because I could imagine this function being quite reusable.

This commit was made using the online interface due to me being away from my normal development computer. If there are formatting issues, I will correct them in a subsequent commit.

* Remove excess whitespace to please linter

Blank lines containing white-space are disallowed.

* Prefer snake_case over camelCase

For the purpose of being able to orally communicate a variable name, etc. it is helpful to have a consistent convention. In Python, this is snake_case. Hence, impose it on the variables `dayOrDays` (which is now `day_or_days`) and `monthOrMonths` (which is now `month_or_months`)

* Use shorter variable names to please linter

The long variable names were causing a line to be too long (121 characters, with the limit being 120). Hence, use the variable names `month_or_s` and `day_or_s` as they seem fairly clear in context, and are also short enough.

* Disable the 'months' countdown when it's 0

When there are no months left, just say how many days are left.

There may be some efficiency concerns as we calculate the months countdown just to disable it! These are likely negligable. In the far future, we might want to consider caching the response every day -- but only if it's worth it.

This code has (still) not been tested.

* Remove trailing whitespace to please linter