add game channel regex function
instead of having to run the :level command, it will look for 'lvl' or 'level' in a message, and then show the level thing instead. it also checks it against what channel it was sent in, so person wont have to specify what base. i tried it on a private server, and it assigned the variables well.