refactor: Account for potential future reactions
In particular:
  - Account for letter reactions of length 1
  - Account for emoji reactions of length other than 1

That is to say, do not use the length of the reaction string to determine whether it is composed of emoji.

To do this, we create a function `react_to_message`. The function takes a message and an iterable, and takes all non-None members of the iterable and proceeds to, in sequence, react to the message with those emojis.

This function then becomes part of the implementation for the existing, and commonly used, `emojify` function