Added the :readme command from scratch. (#159)
* Added the :readme command from scratch.

* Allows ROOT_ROLE_ID users to download a JSON config for the readme. 
* They can then tweak and use that readme file to create the readme on 
any server.
* This is essentially something that allows hotplugging READMEs on any 
server. Solves #147. Cool huh?
* Also, worth mentioning that this is my first bot feature (I've 
literally never touched discord.py before this) and during the 3 days or 
so of dev time that I gave it, I tried my best to write good code that 
flake8 approves and to avoid merge conflicts. If this feature is 
successful, then I'll write the soundboard later if revision and exams 
do not obliterate me.

* No memes in prod

Kthxbai

* Incorporate Empoleon's changes, refactor conditional.

* Squash some juicy computer errors.

* Beep Boop.
* Production is an alias for meme-zone.

* Update readme.json

* Update readme_raw.json