Handling quotes without embeds
For loop was useless, indexing the last element of message.embeds does the same thing.

(with the embed for logging already initialised), if the quote has an embed (ie it is from a bot) then we can set the log description to its description and the log author to its author.

If the quote doesn't, meaning it's from a user, we set the log description to the message content and the log author to the message author.