Updated quotes deletion (#527)
* Updated quotes, fixing #518 #511

When the bot made an embed to send to logs, it tries to mention the users who requested deletion. This failed. str.join is now given an iterable, fixing the error.

* Handling quotes without embeds

For loop was useless, indexing the last element of message.embeds does the same thing.

(with the embed for logging already initialised), if the quote has an embed (ie it is from a bot) then we can set the log description to its description and the log author to its author.

If the quote doesn't, meaning it's from a user, we set the log description to the message content and the log author to the message author.

* Update fun.py

Added two whitespace before inline comments, and cleared empty lines of whitespace. No code changes.

* (Final) bugfixes to quote logging system

After extensive testing I'm pretty sure this works now. Turns out I was getting confused with an embed.author vs a message.author, they aren't the same.

That's been fixed. UnbelievaBoat embeds delete and log, regular text does too.

* Make quote deletion quota an envvar

* Don't try and edit the DB if local debugging

* Don't try and load the quotes to the DB if local debugging

Co-authored-by: Daniel Milnes <thebeanogamer@gmail.com>